completely removed Observers
This change was mostly inspired by this issue, https://github.com/dphfox/Fusion/issues/310

I totally agree with Observers being *completely* redundant, and for the most part unneeded, because most the time when you're creating an Observer it's just to pipe it straight into a `:onBind()` or `:onChange()` call. I've replaced Observers with a function simply called `observe()` which takes the target and the callback as arguments and returns a disconnect function. It adds a set inside the target that holds all of the callbacks and when that target value is evaluate it calls all of those observe tasks, essentially removing the Observer "middleman".

(I know I'm not good at describing these things, you can just look at the changes yourself to see what I mean)