Revert "completely removed Observers"
The observer system I initially designed did not work well with object timeliness, thus lead to catastrophic errors and bugs. Maybe with more time I could have reworked it to work correctly, but that's probably not going to happen.

This is a perfect demonstration of the age old saying, "fuck around and find out."