Update `bindValueToProperty` to require a scope
I decided to just go with the scope route, no need to manually handle the disconnecting of it (that's what the scopes are for anyway). I'm also thinking of applying these changes to the `observe` function (although that's going to be a little tricky, since they're attached to their subject), but it'd be unfitting to make them in this branch, so I'll probably do that on main.